Create NodeTree and refactor Node.
Writing a tree with parent references and mutable fields is tricky
without resorting to unsafe code. Therefore, this change refactors the
Node impl to be a NodeTree, such that all nodes are placed in a vector
in the NodeTree, and Nodes refer to the their children and parent via
indexes.