Make SearchTree a member of Mcts.
The motivation for this is to re-use the search tree in subsequent
moves. When we begin searching a new board position, we update the
SearchTree's root node to be the new board. If the board is one of the
children, then make the child node the root node, otherwise the
SearchTree is reset to start from scratch.