Finish implementing Trainer::train_net().
The only remaining bit was to load the model into a non-autodiff network
for inference, and to return the model so it can be used to play games
against the old network.

Also update the return signature to instead the network wrapped in an
Arc, since that is how the evaluator expects it.